Extract run data preparation module
