+++ apps/server/vibesensor/use_cases/diagnostics/summary_builder.py
@@ -2,10 +2,8 @@
 
 from __future__ import annotations
 
-import math
-from collections.abc import Callable, Sequence
+from collections.abc import Callable
 from dataclasses import dataclass, replace
-from datetime import datetime
 from pathlib import Path
 from statistics import median as _median
 from typing import cast
@@ -19,33 +17,23 @@ from vibesensor.domain import (
     RunSetup,
     RunSuitability,
     Sensor,
-    SpeedProfile,
     SpeedSource,
     TestRun,
 )
 from vibesensor.domain import (
-    DrivingSegment as DomainDrivingSegment,
-)
-from vibesensor.domain import (
     Finding as DomainFinding,
 )
-from vibesensor.domain.snapshots import DrivingPhaseSummary, SpeedProfileSummary
 from vibesensor.domain.test_plan import plan_test_actions
 from vibesensor.domain.vibration_origin import VibrationOrigin
 from vibesensor.report_i18n import normalize_lang
 from vibesensor.shared.boundaries.analysis_payload import (
     AnalysisSummary,
     FindingPayload,
-    PhaseSpeedBreakdownRow,
-    SpeedBreakdownRow,
 )
 from vibesensor.shared.boundaries.diagnostic_case import (
     case_context_from_metadata,
-    speed_profile_from_stats,
 )
 from vibesensor.shared.boundaries.finding import step_payloads_from_plan
-from vibesensor.shared.json_utils import as_float_or_none as _as_float
-from vibesensor.shared.json_utils import i18n_ref
 from vibesensor.shared.run_context import build_summary_warnings
 from vibesensor.shared.time_utils import utc_now_iso
 from vibesensor.shared.types.json_types import JsonObject
@@ -53,35 +41,31 @@ from vibesensor.use_cases.diagnostics._types import (
     AccelStatistics,
     Sample,
 )
-from vibesensor.use_cases.diagnostics.findings import (
-    _build_findings,
-    _phase_speed_breakdown,
-    _sensor_intensity_by_location,
-    _speed_breakdown,
-)
+from vibesensor.use_cases.diagnostics.findings import _build_findings
 from vibesensor.use_cases.diagnostics.helpers import (
     _load_run,
-    _location_label,
-    _locations_connected_throughout_run,
-    _run_noise_baseline_g,
     _validate_required_strength_metrics,
 )
 from vibesensor.use_cases.diagnostics.phase_segmentation import (
     DrivingPhase,
-    PhaseSegment,
 )
 from vibesensor.use_cases.diagnostics.plots import _plot_data
+from vibesensor.use_cases.diagnostics.run_data_preparation import (
+    PreparedRunData,
+    build_domain_driving_segments,
+    build_phase_summary,
+    build_phase_timeline,
+    build_sensor_analysis,
+    prepare_run_data,
+)
 from vibesensor.use_cases.diagnostics.speed_profile_helpers import (
     _speed_stats,
-    _speed_stats_by_phase,
 )
 from vibesensor.use_cases.diagnostics.statistics import (
     _strength_band_key,
     compute_accel_statistics,
     compute_frame_integrity_counts,
     compute_reference_completeness,
-    compute_run_timing,
-    prepare_speed_and_phases,
 )
 from vibesensor.use_cases.diagnostics.summary_serialization import (
     annotate_peaks_with_order_labels,
@@ -90,90 +74,8 @@ from vibesensor.use_cases.diagnostics.summary_serialization import (
 from vibesensor.use_cases.diagnostics.top_cause_selection import select_top_causes
 
 # ═══════════════════════════════════════════════════════════════════════════
-# Phase timeline and driving segments
+# Origin helpers
 # ═══════════════════════════════════════════════════════════════════════════
-
-
-def build_phase_timeline(
-    phase_segments: list[PhaseSegment],
-    findings: Sequence[DomainFinding],
-    *,
-    min_confidence: float,
-) -> list[DrivingPhaseInterval]:
-    """Build a simple phase timeline annotated with finding evidence."""
-    if not phase_segments:
-        return []
-
-    # NOTE: has_fault_evidence is always False because phases_detected is not
-    # preserved on the domain Finding (only cruise_fraction survives the
-    # payload→domain decode).  Keeping the field for schema stability.
-    return [
-        DrivingPhaseInterval(
-            phase=segment.phase,
-            start_t_s=None if math.isnan(segment.start_t_s) else segment.start_t_s,
-            end_t_s=None if math.isnan(segment.end_t_s) else segment.end_t_s,
-            speed_min_kmh=segment.speed_min_kmh,
-            speed_max_kmh=segment.speed_max_kmh,
-            has_fault_evidence=False,
-        )
-        for segment in phase_segments
-    ]
-
-
-def build_domain_driving_segments(
-    phase_segments: list[PhaseSegment],
-) -> tuple[DomainDrivingSegment, ...]:
-    return tuple(
-        DomainDrivingSegment(
-            phase=segment.phase,
-            start_idx=segment.start_idx,
-            end_idx=segment.end_idx,
-            start_t_s=(
-                None
-                if isinstance(segment.start_t_s, float) and math.isnan(segment.start_t_s)
-                else segment.start_t_s
-            ),
-            end_t_s=(
-                None
-                if isinstance(segment.end_t_s, float) and math.isnan(segment.end_t_s)
-                else segment.end_t_s
-            ),
-            speed_min_kmh=segment.speed_min_kmh,
-            speed_max_kmh=segment.speed_max_kmh,
-            sample_count=segment.sample_count,
-        )
-        for segment in phase_segments
-    )
-
-
-# ═══════════════════════════════════════════════════════════════════════════
-# Sensor analysis and summary helpers
-# ═══════════════════════════════════════════════════════════════════════════
-
-
-def build_sensor_analysis(
-    *,
-    samples: list[Sample],
-    language: str,
-    per_sample_phases: list[DrivingPhase],
-) -> tuple[list[str], set[str], list[LocationIntensitySummary]]:
-    """Build sensor location lists and intensity rows from analysed samples."""
-    sensor_locations = sorted(
-        {
-            label
-            for sample in samples
-            if isinstance(sample, dict) and (label := _location_label(sample, lang=language))
-        },
-    )
-    connected_locations = _locations_connected_throughout_run(samples, lang=language)
-    sensor_intensity_by_location = _sensor_intensity_by_location(
-        samples,
-        include_locations=set(sensor_locations),
-        lang=language,
-        connected_locations=connected_locations,
-        per_sample_phases=per_sample_phases,
-    )
-    return sensor_locations, connected_locations, sensor_intensity_by_location
 
 
 def summarize_origin(
@@ -181,106 +83,6 @@ def summarize_origin(
 ) -> VibrationOrigin | None:
     """Return the most-likely origin as a domain value object."""
     return VibrationOrigin.from_ranked_findings(findings)
-
-
-# ═══════════════════════════════════════════════════════════════════════════
-# Main orchestration
-# ═══════════════════════════════════════════════════════════════════════════
-
-
-@dataclass(frozen=True)
-class PreparedRunData:
-    """Input coordinator: shared timing, speed, and phase context for summary generation.
-
-    Retained as the canonical input coordinator for the analysis pipeline.
-    Computed once by :func:`prepare_run_data` and consumed by
-    :func:`build_findings_bundle`, :func:`build_run_suitability_bundle`,
-    and :class:`RunAnalysis`.
-    """
-
-    run_id: str
-    start_ts: datetime | None
-    end_ts: datetime | None
-    duration_s: float
-    raw_sample_rate_hz: float | None
-    speed_values: list[float]
-    speed_non_null_pct: float
-    speed_sufficient: bool
-    per_sample_phases: list[DrivingPhase]
-    phase_segments: list[PhaseSegment]
-    run_noise_baseline_g: float | None
-    speed_profile: SpeedProfile
-    speed_stats_by_phase: dict[str, SpeedProfileSummary]
-    speed_breakdown: list[SpeedBreakdownRow]
-    speed_breakdown_skipped_reason: JsonObject | None
-    phase_speed_breakdown: list[PhaseSpeedBreakdownRow]
-
-    # -- derived convenience ------------------------------------------------
-
-    @property
-    def is_steady_speed(self) -> bool:
-        """Whether the run had steady speed (relevant to confidence scoring)."""
-        steady: bool = self.speed_profile.steady_speed
-        return steady
-
-    @property
-    def speed_stddev_kmh(self) -> float | None:
-        return self.speed_profile.stddev_kmh if self.speed_values else None
-
-
-def prepare_run_data(
-    metadata: JsonObject,
-    samples: list[Sample],
-    *,
-    file_name: str,
-) -> PreparedRunData:
-    """Prepare shared timing, speed, and phase context for summary generation."""
-    run_id, start_ts, end_ts, duration_s = compute_run_timing(metadata, samples, file_name)
-    (
-        speed_values,
-        speed_stats,
-        speed_non_null_pct,
-        speed_sufficient,
-        per_sample_phases,
-        phase_segments,
-    ) = prepare_speed_and_phases(samples)
-    run_noise_baseline_g = _run_noise_baseline_g(samples)
-    speed_breakdown = _speed_breakdown(samples) if speed_sufficient else []
-    speed_breakdown_skipped_reason: JsonObject | None = None
-    if not speed_sufficient:
-        speed_breakdown_skipped_reason = i18n_ref(
-            "SPEED_DATA_MISSING_OR_INSUFFICIENT_SPEED_BINNED_AND",
-        )
-    phase_info = build_phase_summary(phase_segments)
-
-    return PreparedRunData(
-        run_id=run_id,
-        start_ts=start_ts,
-        end_ts=end_ts,
-        duration_s=duration_s,
-        raw_sample_rate_hz=_as_float(metadata.get("raw_sample_rate_hz")),
-        speed_values=speed_values,
-        speed_non_null_pct=speed_non_null_pct,
-        speed_sufficient=speed_sufficient,
-        per_sample_phases=per_sample_phases,
-        phase_segments=phase_segments,
-        run_noise_baseline_g=run_noise_baseline_g,
-        speed_profile=speed_profile_from_stats(
-            speed_stats,
-            phase_info,
-        ),
-        speed_stats_by_phase=_speed_stats_by_phase(samples, per_sample_phases),
-        speed_breakdown=speed_breakdown,
-        speed_breakdown_skipped_reason=speed_breakdown_skipped_reason,
-        phase_speed_breakdown=_phase_speed_breakdown(samples, per_sample_phases),
-    )
-
-
-def build_phase_summary(phase_segments: list[PhaseSegment]) -> DrivingPhaseSummary:
-    """Small wrapper to keep summary-building imports localized."""
-    from vibesensor.use_cases.diagnostics.phase_segmentation import phase_summary
-
-    return phase_summary(phase_segments)
 
 
 def build_findings_bundle(
